requirements.txt
phi
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
pycountry
googlesearch-python
